requirements.txt
# Core Django Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-redis==5.4.0

# Database & Storage
psycopg2-binary==2.9.9
pyarrow==13.0.0
pandas==2.1.3
numpy==1.24.3

# Task Queue and Background Processing (NON-NEGOTIABLE)
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Data Analysis Libraries (NON-NEGOTIABLE - NO pandas-ai in production)
lifelines==0.27.8
matplotlib==3.8.2
seaborn==0.13.0

# File Processing
openpyxl==3.1.2

# AI Integration
langchain==0.1.0
google-generativeai==0.3.2
tiktoken==0.5.2

# Report Generation
python-docx==1.1.0
Pillow==10.1.0

# Security and Utilities
django-cleanup==8.1.0
bleach==6.1.0
psutil==5.9.6
cryptography==41.0.7
whitenoise==6.6.0
gunicorn==21.2.0

# Development and Testing
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Testing
factory-boy==3.3.0
faker==20.1.0
